Move objectives to their own file (from utils)
