+++ nengo_dl/__init__.py
@@ -32,7 +32,7 @@ from nengo_dl import (
     transform_builders)
 
 # import into top-level namespace
-from nengo_dl import dists
+from nengo_dl import dists, objectives
 from nengo_dl.simulator import Simulator
 from nengo_dl.tensor_node import TensorNode, tensor_layer, reshaped
 from nengo_dl.config import configure_settings
